bin/beomon_cli.py:
Usage: beomon_cli.py [options] [node1] [node2] ...

Beomon command line interface.
This program displays the status of the cluster in plain text.  When ran
with no arguements an overall status is printed.  With an arguement of a node
name (e.g. 'head0a' or 'n17' or '17') the status of that node is printed.